explain-validator-requirements.txt
sqlparse
snowflake-connector-python
duckdb
pathlib
jinja2
python-dotenv